[BE] feat : 마일리지 추가

+++ backend/board/src/main/java/com/newbie/board/generalBoard/controller/GeneralBoardCommentController.java
@@ -29,7 +29,7 @@ public class GeneralBoardCommentController {
     @PostMapping("/{boardId}")
     public ResponseEntity<GeneralBoardCommentResponseDto> createComment(
             @RequestBody @Parameter(description = "boardId, content") GeneralBoardCommentRequestDto requestDto,
-            @RequestParam Long userId) {
+            @RequestParam int userId) {
         GeneralBoardCommentResponseDto createdComment = commentService.createComment(requestDto, userId);
         return ResponseEntity.status(HttpStatus.CREATED).body(createdComment);
     }
